エンディング全体の設計
