+++ Assets/MyScripts/Ending.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 767670ef05c86424fb8aed2f39b399ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
