+++ Assets/MyScripts/Ending/EndingFail.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90633d12484fdfb4b8dea33b64b8dfef
+guid: df623e31743b68046bdc03f0eedfefba
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
