+++ Assets/MyScripts/Ending/EndingSuccess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc495ca02a4c78d4f891c40333cbd919
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
