+++ Assets/Scenes/sakoda.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:78b1231ecd0f1984ed3d3dc9018c8476e5132e8b1ad6ddcaa38e24839c6e4cb6
-size 22255
+oid sha256:bda1a7646382300dbcc8ad63e9a546e0b3a9086714e93e1b9c7e5da196f14cc6
+size 38386
